Create Script 1.ts
This is a TypeScript function that takes in four parameters:

workbook: an ExcelScript.Workbook object representing an Excel workbook.
sheetName: a string representing the name of the worksheet that contains the chart. It has a default value of "Sheet1" if not provided.
chartName: a string representing the name of the chart. It has a default value of "Chart 2" if not provided.
chartWidth: a number representing the width of the chart image in pixels. It has a default value of 550 if not provided.
The function first retrieves the Excel chart object using the getWorksheet() and getChart() methods of the workbook object, passing in the sheetName and chartName parameters, respectively.

Then, it calls the getImage() method on the chart object, passing in the chartWidth parameter to get a PNG image of the chart with the specified width.

Finally, it creates a data URI for the chart image by prefixing the base64-encoded image with the data:image/png;base64, string and returns the result as a string.

The function can be used to generate a PNG image of an Excel chart in a specified worksheet and with a specified size. The resulting image can be used for various purposes, such as inserting it into a Word or PowerPoint document or displaying it on a webpage.