Create Weekly Traveler Report.ts
This is a TypeScript function that takes in three parameters:

workbook: an ExcelScript.Workbook object representing an Excel workbook.
worksheetToKeepName: a string representing the name of the worksheet that should be kept visible.
visibilityType: a string representing the visibility type that should be applied to the other worksheets in the workbook. This can be either "hidden" or "veryHidden".
The function starts by getting all the worksheets in the workbook using the getWorksheets() method of the workbook object. It then unprotects all the worksheets in the workbook using the unprotect() method of the getProtection() object for each worksheet in the worksheet collection.

Next, the function unhides all the worksheets in the workbook using the setVisibility() method of each worksheet object and passing in the ExcelScript.SheetVisibility.visible constant as an argument.

Then, the function gets the active worksheet using the getActiveWorksheet() method of the workbook object and gets the used range in the active worksheet using the getUsedRange() method of the worksheet object.

After that, the function creates a variable wsArr to hold the worksheet collection of the workbook.

Next, the function loops through all the worksheets in the wsArr worksheet collection and checks if the name of the worksheet is the same as the worksheet to retain. If the worksheet is not the worksheet to retain, the function uses a switch statement to check the visibilityType argument and set the visibilityTypeScript variable to the corresponding ExcelScript.SheetVisibility constant. If the visibilityType argument is not valid, the function logs an error message to the console and sets the visibilityTypeScript variable to ExcelScript.SheetVisibility.hidden.

Finally, the function sets the visibility of the worksheet to the visibilityTypeScript value using the setVisibility() method of the worksheet object.

In summary, the function can be used to retain one worksheet and hide or very-hide all other worksheets in an Excel workbook. This can be useful for protecting sensitive data or making specific worksheets easily accessible.